+++ appstream-center-20210220/alibabacloud_appstream_center20210220/client.py
@@ -44,17 +44,24 @@ class Client(OpenApiClient):
 
     def find_idp_list_by_login_identifier_with_options(
         self,
-        request: appstream_center_20210220_models.FindIdpListByLoginIdentifierRequest,
+        tmp_req: appstream_center_20210220_models.FindIdpListByLoginIdentifierRequest,
         runtime: util_models.RuntimeOptions,
     ) -> appstream_center_20210220_models.FindIdpListByLoginIdentifierResponse:
         """
         @summary 身份认证查询接口
         
-        @param request: FindIdpListByLoginIdentifierRequest
+        @param tmp_req: FindIdpListByLoginIdentifierRequest
         @param runtime: runtime options for this request RuntimeOptions
         @return: FindIdpListByLoginIdentifierResponse
         """
-        UtilClient.validate_model(request)
+        UtilClient.validate_model(tmp_req)
+        request = appstream_center_20210220_models.FindIdpListByLoginIdentifierShrinkRequest()
+        OpenApiUtilClient.convert(tmp_req, request)
+        if not UtilClient.is_unset(tmp_req.available_features):
+            request.available_features_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.available_features, 'AvailableFeatures', 'json')
+        query = {}
+        if not UtilClient.is_unset(request.available_features_shrink):
+            query['AvailableFeatures'] = request.available_features_shrink
         body = {}
         if not UtilClient.is_unset(request.client_channel):
             body['ClientChannel'] = request.client_channel
@@ -73,6 +80,7 @@ class Client(OpenApiClient):
         if not UtilClient.is_unset(request.uuid):
             body['Uuid'] = request.uuid
         req = open_api_models.OpenApiRequest(
+            query=OpenApiUtilClient.query(query),
             body=OpenApiUtilClient.parse_to_map(body)
         )
         params = open_api_models.Params(
@@ -93,17 +101,24 @@ class Client(OpenApiClient):
 
     async def find_idp_list_by_login_identifier_with_options_async(
         self,
-        request: appstream_center_20210220_models.FindIdpListByLoginIdentifierRequest,
+        tmp_req: appstream_center_20210220_models.FindIdpListByLoginIdentifierRequest,
         runtime: util_models.RuntimeOptions,
     ) -> appstream_center_20210220_models.FindIdpListByLoginIdentifierResponse:
         """
         @summary 身份认证查询接口
         
-        @param request: FindIdpListByLoginIdentifierRequest
+        @param tmp_req: FindIdpListByLoginIdentifierRequest
         @param runtime: runtime options for this request RuntimeOptions
         @return: FindIdpListByLoginIdentifierResponse
         """
-        UtilClient.validate_model(request)
+        UtilClient.validate_model(tmp_req)
+        request = appstream_center_20210220_models.FindIdpListByLoginIdentifierShrinkRequest()
+        OpenApiUtilClient.convert(tmp_req, request)
+        if not UtilClient.is_unset(tmp_req.available_features):
+            request.available_features_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.available_features, 'AvailableFeatures', 'json')
+        query = {}
+        if not UtilClient.is_unset(request.available_features_shrink):
+            query['AvailableFeatures'] = request.available_features_shrink
         body = {}
         if not UtilClient.is_unset(request.client_channel):
             body['ClientChannel'] = request.client_channel
@@ -122,6 +137,7 @@ class Client(OpenApiClient):
         if not UtilClient.is_unset(request.uuid):
             body['Uuid'] = request.uuid
         req = open_api_models.OpenApiRequest(
+            query=OpenApiUtilClient.query(query),
             body=OpenApiUtilClient.parse_to_map(body)
         )
         params = open_api_models.Params(
@@ -168,20 +184,26 @@ class Client(OpenApiClient):
 
     def get_login_token_with_options(
         self,
-        request: appstream_center_20210220_models.GetLoginTokenRequest,
+        tmp_req: appstream_center_20210220_models.GetLoginTokenRequest,
         runtime: util_models.RuntimeOptions,
     ) -> appstream_center_20210220_models.GetLoginTokenResponse:
         """
         @summary GetLoginToken
         
-        @param request: GetLoginTokenRequest
+        @param tmp_req: GetLoginTokenRequest
         @param runtime: runtime options for this request RuntimeOptions
         @return: GetLoginTokenResponse
         """
-        UtilClient.validate_model(request)
+        UtilClient.validate_model(tmp_req)
+        request = appstream_center_20210220_models.GetLoginTokenShrinkRequest()
+        OpenApiUtilClient.convert(tmp_req, request)
+        if not UtilClient.is_unset(tmp_req.available_features):
+            request.available_features_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.available_features, 'AvailableFeatures', 'json')
         query = {}
         if not UtilClient.is_unset(request.authentication_code):
             query['AuthenticationCode'] = request.authentication_code
+        if not UtilClient.is_unset(request.available_features_shrink):
+            query['AvailableFeatures'] = request.available_features_shrink
         if not UtilClient.is_unset(request.client_id):
             query['ClientId'] = request.client_id
         if not UtilClient.is_unset(request.client_os):
@@ -216,6 +238,8 @@ class Client(OpenApiClient):
             query['LoginIdentifier'] = request.login_identifier
         if not UtilClient.is_unset(request.login_name):
             query['LoginName'] = request.login_name
+        if not UtilClient.is_unset(request.mfa_type):
+            query['MfaType'] = request.mfa_type
         if not UtilClient.is_unset(request.network_type):
             query['NetworkType'] = request.network_type
         if not UtilClient.is_unset(request.new_password):
@@ -265,20 +289,26 @@ class Client(OpenApiClient):
 
     async def get_login_token_with_options_async(
         self,
-        request: appstream_center_20210220_models.GetLoginTokenRequest,
+        tmp_req: appstream_center_20210220_models.GetLoginTokenRequest,
         runtime: util_models.RuntimeOptions,
     ) -> appstream_center_20210220_models.GetLoginTokenResponse:
         """
         @summary GetLoginToken
         
-        @param request: GetLoginTokenRequest
+        @param tmp_req: GetLoginTokenRequest
         @param runtime: runtime options for this request RuntimeOptions
         @return: GetLoginTokenResponse
         """
-        UtilClient.validate_model(request)
+        UtilClient.validate_model(tmp_req)
+        request = appstream_center_20210220_models.GetLoginTokenShrinkRequest()
+        OpenApiUtilClient.convert(tmp_req, request)
+        if not UtilClient.is_unset(tmp_req.available_features):
+            request.available_features_shrink = OpenApiUtilClient.array_to_string_with_specified_style(tmp_req.available_features, 'AvailableFeatures', 'json')
         query = {}
         if not UtilClient.is_unset(request.authentication_code):
             query['AuthenticationCode'] = request.authentication_code
+        if not UtilClient.is_unset(request.available_features_shrink):
+            query['AvailableFeatures'] = request.available_features_shrink
         if not UtilClient.is_unset(request.client_id):
             query['ClientId'] = request.client_id
         if not UtilClient.is_unset(request.client_os):
@@ -313,6 +343,8 @@ class Client(OpenApiClient):
             query['LoginIdentifier'] = request.login_identifier
         if not UtilClient.is_unset(request.login_name):
             query['LoginName'] = request.login_name
+        if not UtilClient.is_unset(request.mfa_type):
+            query['MfaType'] = request.mfa_type
         if not UtilClient.is_unset(request.network_type):
             query['NetworkType'] = request.network_type
         if not UtilClient.is_unset(request.new_password):